Move value objects in common to shared package and add methods to validate them
