+++ src/main/java/com/hexagonal/shop/cart/application/confirm/ConfirmPurchase.java
@@ -3,10 +3,10 @@ package com.hexagonal.shop.cart.application.confirm;
 import com.hexagonal.shop.shared.domain.bus.event.EventBus;
 import com.hexagonal.shop.cart.domain.Cart;
 import com.hexagonal.shop.cart.domain.CartRepository;
-import com.hexagonal.shop.cart.domain.valueobject.Address;
-import com.hexagonal.shop.cart.domain.valueobject.CartId;
-import com.hexagonal.shop.cart.domain.valueobject.DiscountCode;
-import com.hexagonal.shop.cart.domain.valueobject.Email;
+import com.hexagonal.shop.shared.domain.valueobject.Address;
+import com.hexagonal.shop.shared.domain.valueobject.CartId;
+import com.hexagonal.shop.shared.domain.valueobject.DiscountCode;
+import com.hexagonal.shop.shared.domain.valueobject.Email;
 
 import java.util.Optional;
