--- src/main/java/com/hexagonal/shop/cart/domain/valueobject/Address.java
@@ -1,9 +1,0 @@
-package com.hexagonal.shop.cart.domain.valueobject;
-
-import com.hexagonal.shop.shared.domain.valueobject.StringValueObject;
-
-public class Address extends StringValueObject {
-    public Address(String value) {
-        super(value);
-    }
-}
