--- src/main/java/com/hexagonal/shop/cart/domain/valueobject/CartId.java
@@ -1,10 +1,0 @@
-package com.hexagonal.shop.cart.domain.valueobject;
-
-import com.hexagonal.shop.shared.domain.valueobject.Identifier;
-
-public class CartId extends Identifier {
-
-    public CartId(String id) {
-        super(id);
-    }
-}
