--- src/main/java/com/hexagonal/shop/cart/domain/valueobject/DiscountCode.java
@@ -1,9 +1,0 @@
-package com.hexagonal.shop.cart.domain.valueobject;
-
-import com.hexagonal.shop.shared.domain.valueobject.StringValueObject;
-
-public class DiscountCode extends StringValueObject {
-    public DiscountCode(String value) {
-        super(value);
-    }
-}
