--- src/main/java/com/hexagonal/shop/cart/domain/valueobject/Email.java
@@ -1,9 +1,0 @@
-package com.hexagonal.shop.cart.domain.valueobject;
-
-import com.hexagonal.shop.shared.domain.valueobject.StringValueObject;
-
-public class Email extends StringValueObject {
-    public Email(String value) {
-        super(value);
-    }
-}
