--- src/main/java/com/hexagonal/shop/cart/domain/valueobject/ProductId.java
@@ -1,9 +1,0 @@
-package com.hexagonal.shop.cart.domain.valueobject;
-
-import com.hexagonal.shop.shared.domain.valueobject.Identifier;
-
-public class ProductId extends Identifier {
-    public ProductId(String value) {
-        super(value);
-    }
-}
