+++ src/main/java/com/hexagonal/shop/order/domain/Order.java
@@ -2,10 +2,10 @@ package com.hexagonal.shop.order.domain;
 
 import com.hexagonal.shop.shared.domain.AggregateRoot;
 import com.hexagonal.shop.cart.domain.AmountProducts;
-import com.hexagonal.shop.cart.domain.valueobject.Address;
-import com.hexagonal.shop.cart.domain.valueobject.DiscountCode;
-import com.hexagonal.shop.cart.domain.valueobject.Email;
-import com.hexagonal.shop.cart.domain.valueobject.ProductId;
+import com.hexagonal.shop.shared.domain.valueobject.Address;
+import com.hexagonal.shop.shared.domain.valueobject.DiscountCode;
+import com.hexagonal.shop.shared.domain.valueobject.Email;
+import com.hexagonal.shop.shared.domain.valueobject.ProductId;
 import com.hexagonal.shop.shared.domain.product.Product;
 
 import java.util.List;
