+++ src/main/java/com/hexagonal/shop/order/domain/valueobject/Price.java
@@ -3,8 +3,14 @@ package com.hexagonal.shop.order.domain.valueobject;
 import com.hexagonal.shop.shared.domain.valueobject.IntValueObject;
 
 public class Price extends IntValueObject {
+
     public Price(Integer value) {
         super(value);
-        // FIXME: 9/19/2021 ensure not negative value
+    }
+
+    @Override
+    protected void ensureValid(Integer value) {
+        if (value <= 0)
+            throw new IllegalArgumentException("Price cannot be less than 1");
     }
 }
