+++ src/main/java/com/hexagonal/shop/order/domain/valueobject/ProductName.java
@@ -3,7 +3,14 @@ package com.hexagonal.shop.order.domain.valueobject;
 import com.hexagonal.shop.shared.domain.valueobject.StringValueObject;
 
 public class ProductName extends StringValueObject {
+
     public ProductName(String value) {
         super(value);
+    }
+
+    @Override
+    protected void ensureValid(String value) {
+        if (value.isEmpty())
+            throw new IllegalArgumentException("Product Name cannot be empty");
     }
 }
