+++ src/main/java/com/hexagonal/shop/shared/domain/product/ProductRepository.java
@@ -1,6 +1,6 @@
 package com.hexagonal.shop.shared.domain.product;
 
-import com.hexagonal.shop.cart.domain.valueobject.ProductId;
+import com.hexagonal.shop.shared.domain.valueobject.ProductId;
 
 import java.util.List;
 import java.util.Set;
