+++ src/main/java/com/hexagonal/shop/shared/domain/valueobject/Address.java
@@ -1,0 +1,8 @@
+package com.hexagonal.shop.shared.domain.valueobject;
+
+public class Address extends StringValueObject {
+
+    public Address(String value) {
+        super(value);
+    }
+}
