+++ src/main/java/com/hexagonal/shop/shared/domain/valueobject/CartId.java
@@ -1,0 +1,8 @@
+package com.hexagonal.shop.shared.domain.valueobject;
+
+public class CartId extends Identifier {
+
+    public CartId(String id) {
+        super(id);
+    }
+}
