+++ src/main/java/com/hexagonal/shop/shared/domain/valueobject/ProductId.java
@@ -1,0 +1,8 @@
+package com.hexagonal.shop.shared.domain.valueobject;
+
+public class ProductId extends Identifier {
+
+    public ProductId(String value) {
+        super(value);
+    }
+}
