+++ src/main/java/com/hexagonal/shop/shared/domain/valueobject/StringValueObject.java
@@ -3,14 +3,19 @@ package com.hexagonal.shop.shared.domain.valueobject;
 import java.util.Objects;
 
 public abstract class StringValueObject {
+
     private String value;
 
     public StringValueObject(String value) {
+        ensureValid(value);
         this.value = value;
     }
 
     public String value() {
         return value;
+    }
+
+    protected void ensureValid(String value) {
     }
 
     @Override
