+++ src/main/java/com/hexagonal/shop/shared/infraestructure/product/HibernateProductRepository.java
@@ -1,6 +1,6 @@
 package com.hexagonal.shop.shared.infraestructure.product;
 
-import com.hexagonal.shop.shared.domain.valueobject.ProductEntity;
+import com.hexagonal.shop.shared.domain.product.ProductEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
